+++ cli/csv-to-xlsx.types.ts
@@ -1,0 +1,5 @@
+export type CLIParameters = {
+  inputDir: string;
+  outputDir: string;
+  sheetName: string;
+};